requirements.txt
aiogram
python-dotenv
pyTelegramBotAPI